==> notebooks/01_filter_amex_complaints.ipynb <==
!pip install --upgrade pip

!pip install pandas numpy matplotlib seaborn nltk scikit-learn vaderSentiment

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import nltk
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import classification_report

%matplotlib inline

nltk.download('vader_lexicon')


df = pd.read_csv("../data/intermediate data/complaints.csv", nrows=1)
print(df.columns.tolist())



input_file = "../data/intermediate data/complaints.csv"
output_file = "../data/intermediate data/amex_credit_card_complaints.csv"

product_filter = "credit card"
company_filter = "american express"

chunk_size = 100_000
filtered_chunks = []

for chunk in pd.read_csv(input_file, chunksize=chunk_size, low_memory=False):

    chunk['Product'] = chunk['Product'].str.lower()
    chunk['Company'] = chunk['Company'].str.lower()

    filtered = chunk[
        (chunk['Product'] == product_filter) &
        (chunk['Company'].str.contains(company_filter))
    ]
    filtered_chunks.append(filtered)

final_df = pd.concat(filtered_chunks, ignore_index=True)
print(final_df.shape)
final_df.to_csv(output_file, index=False)

print(f"Done! Filtered rows: {len(final_df)}")
print(f"Saved to: {output_file}")
