==> notebooks/02_cleaning_eda.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

pd.set_option('display.max_columns', 100)
sns.set(style='whitegrid')

df = pd.read_csv("../data/intermediate data/amex_credit_card_complaints.csv")

print("Dataset Loaded!")
print("Shape:", df.shape)
df.head()


df.columns = df.columns.str.lower().str.replace(" ", "_")

df['date_received'] = pd.to_datetime(df['date_received'])
df['date_sent_to_company'] = pd.to_datetime(df['date_sent_to_company'])

print("Column names cleaned & date columns parsed.")


missing = df.isnull().mean().sort_values(ascending=False)
missing = missing[missing > 0]
print("Columns with missing values:\n", missing)

df = df.dropna(subset=['consumer_complaint_narrative'])

df['consumer_disputed?'] = df['consumer_disputed?'].fillna("Unknown")
df['timely_response?'] = df['timely_response?'].fillna("Unknown")
df['tags'] = df['tags'].fillna("Not Provided")
df['consumer_consent_provided?'] = df['consumer_consent_provided?'].fillna("Not Provided")
df['company_public_response'] = df['company_public_response'].fillna("No response")

print("Missing values handled.")


df.drop(columns=['company_public_response'], inplace=True)

df = df.drop_duplicates(subset=['consumer_complaint_narrative', 'issue'])

df = df[df['consumer_complaint_narrative'].str.len() > 30]

df = df.reset_index(drop=True)

print("Removed duplicates and too-short complaints.")


print("Number of rows:", df.shape[0])
print("Unique issues:", df['issue'].nunique())
print("Unique submission channels:", df['submitted_via'].unique())
print("Date range:", df['date_received'].min(), "to", df['date_received'].max())


df.to_csv("../data/intermediate data/amex_cleaned_complaints.csv", index=False)
print("Cleaned dataset saved to: ../data/intermediate data/amex_cleaned_complaints.csv")
