==> notebooks/03_nlp_sentiment.ipynb <==
import pandas as pd

df = pd.read_csv("../data/intermediate data/amex_cleaned_complaints.csv")

print("Rows:", len(df))
print("Sample complaint:\n", df['consumer_complaint_narrative'].iloc[0])


import re

def clean_text(text):
    text = text.lower()
    text = re.sub(r'http\S+|www\S+', '', text)           
    text = re.sub(r'[^a-z\s]', '', text)                  
    text = re.sub(r'\s+', ' ', text).strip()            
    return text

df['clean_text'] = df['consumer_complaint_narrative'].apply(clean_text)
print("Text cleaned.")


import nltk
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer

nltk.download('vader_lexicon')

analyzer = SentimentIntensityAnalyzer()

df['vader_sentiment'] = df['clean_text'].apply(lambda x: analyzer.polarity_scores(x)['compound'])

df['sentiment_label'] = df['vader_sentiment'].apply(
    lambda x: 'positive' if x > 0.2 else 'negative' if x < -0.2 else 'neutral'
)

print("Sentiment analysis complete.")
print("Sentiment counts:\n", df['sentiment_label'].value_counts())


df.to_csv("../data/amex_complaints_sentiment.csv", index=False)
print("Saved sentiment-scored file to: ../data/amex_complaints_sentiment.csv")
