==> notebooks/07_complaints_clean.ipynb <==
import pandas as pd
import html

df = pd.read_csv("../data/amex_complaints_sentiment.csv")

text_cols = ['consumer_complaint_narrative', 'clean_text']
for col in text_cols:
    df[col] = df[col].fillna("")
    df[col] = df[col].apply(lambda x: str(x).replace('\n', ' ').replace('\r', ' '))
    df[col] = df[col].apply(lambda x: x.replace('"', "'"))
    df[col] = df[col].apply(lambda x: html.unescape(x))
    
df.columns = (
    df.columns.str.lower()
              .str.strip()
              .str.replace(' ', '_')
              .str.replace(r'\W', '', regex=True)  
)


df.to_csv("../outputs/amex_complaints_clean.csv", index=False, quoting=1)
print('Data cleaned and saved')
