==> notebooks/09_bigquery_integration.ipynb <==
!pip install --upgrade google-cloud-bigquery pandas


!pip install db-dtypes


from google.cloud import bigquery

project_id = "amex-analytics-project"
dataset_id = "amex_data"
table_id = "amex_complaints"

client = bigquery.Client(project=project_id)


query = f"""
SELECT * 
FROM `{project_id}.{dataset_id}.{table_id}`
LIMIT 10
"""
df = client.query(query).to_dataframe()
df.head()


query = f"""
SELECT COUNT(*) as total_complaints
FROM `{project_id}.{dataset_id}.{table_id}`
"""
client.query(query).to_dataframe()


query = f"""
SELECT issue, COUNT(*) as complaint_count
FROM `{project_id}.{dataset_id}.{table_id}`
GROUP BY issue
ORDER BY complaint_count DESC
LIMIT 10
"""
client.query(query).to_dataframe()


query = f"""
SELECT submitted_via, COUNT(*) as count
FROM `{project_id}.{dataset_id}.{table_id}`
GROUP BY submitted_via
ORDER BY count DESC
"""
client.query(query).to_dataframe()


query = f"""
SELECT FORMAT_DATE('%Y-%m', date_received) as month,
       COUNT(*) as total_complaints
FROM `{project_id}.{dataset_id}.{table_id}`
GROUP BY month
ORDER BY month
"""
client.query(query).to_dataframe()


query = f"""
SELECT sentiment_label, COUNT(*) as count
FROM `{project_id}.{dataset_id}.{table_id}`
GROUP BY sentiment_label
ORDER BY count DESC
"""
client.query(query).to_dataframe()


query = f"""
WITH top_issues AS (
    SELECT issue
    FROM `{project_id}.{dataset_id}.{table_id}`
    GROUP BY issue
    ORDER BY COUNT(*) DESC
    LIMIT 5
)
SELECT issue, sentiment_label, COUNT(*) as count
FROM `{project_id}.{dataset_id}.{table_id}`
WHERE issue IN (SELECT issue FROM top_issues)
GROUP BY issue, sentiment_label
ORDER BY issue, count DESC
"""
client.query(query).to_dataframe()


query = f"""
SELECT sentiment_label, timely_response, COUNT(*) as count
FROM `{project_id}.{dataset_id}.{table_id}`
GROUP BY sentiment_label, timely_response
ORDER BY sentiment_label, count DESC
"""
df = client.query(query).to_dataframe()
df.head()
